refactor(core): simplify source layout for contributors
Final structural cleanup before v0.4.3 ships. No behavior, schema, or
public API changes — only file moves, internal renames, and import paths.

tools/ renamed for a new-contributor read:
- auth/: detector→detect, explorer→explore, block-gap→gaps,
  surface-analyzer→surface, oauth-providers→providers,
  user-providers→custom-providers (apply.ts unchanged)
- explorers/: playwright-explorer→playwright, cypress-explorer→cypress,
  explorer.interface→types
- repo/: scanner→scan, framework-detector→detect-framework
- scoring/: gap-engine→gaps

packages/mcp:
- compact-analyze-payload(.ts/.test.ts) → summarize-analyze-result.ts
  + __tests__/summarize-analyze-result.test.ts
- internal function buildCompactAnalyzePayload → summarizeAnalyzeResult
  (internal symbol only; no public export changed)

Root cleanup:
- removed .scan-state/.gitkeep and output/.gitkeep (runtime creates these
  via mkdir(..., { recursive: true }) in state-manager.ts and the JSON/
  Markdown reporters)
- simplified .gitignore: .scan-state/ and output/ ignored outright, no
  .gitkeep exception needed

All internal imports, dynamic imports, test scripts, and JSDoc @module
tags updated. Public exports in packages/core/src/index.ts unchanged
(detectAuth, exploreAuth, scanRepo, computeAutomationMaturity,
validateStorageState, etc. all still resolve via the new paths).

Verified: npm run build, npm test (66 core + 6 mcp), npm run
test:integration (2 ran, 2 skipped network-gated), and dry-run
publishes for @qulib/core and @qulib/mcp all green.
