--- output/.gitkeep
@@ -1,2 +1,0 @@
-This folder is generated by `qulib analyze`. It contains report.json, report.md,
-and a `generated/` subfolder with test scaffolds. Run `qulib clean` to reset.
