+++ packages/core/src/tools/auth/detect.ts
@@ -4,7 +4,7 @@ import type { Page } from '@playwright/test';
 import type { AuthPath, DetectedAuth } from '../../schemas/config.schema.js';
 import type { AnalyzeProgressSink } from '../../harness/progress-log.js';
 import { launchBrowser } from '../explorers/browser.js';
-import { BUILT_IN_OAUTH_PROVIDERS } from './oauth-providers.js';
+import { BUILT_IN_OAUTH_PROVIDERS } from './providers.js';
 
 export type StorageStateInvalidReason =
   | 'missing-file'
