+++ packages/core/src/tools/auth/providers.ts
