+++ packages/core/src/tools/explorers/cypress.ts
@@ -1,4 +1,4 @@
-import type { AppExplorer } from './explorer.interface.js';
+import type { AppExplorer } from './types.js';
 import type { HarnessConfig } from '../../schemas/config.schema.js';
 import type { RouteInventory } from '../../schemas/route-inventory.schema.js';
 import type { RunArtifactsOptions } from '../../harness/run-options.js';
