+++ packages/core/src/tools/explorers/playwright.ts
@@ -1,7 +1,7 @@
 import type { BrowserContext } from '@playwright/test';
 import { launchBrowser } from './browser.js';
 import { AxeBuilder } from '@axe-core/playwright';
-import type { AppExplorer } from './explorer.interface.js';
+import type { AppExplorer } from './types.js';
 import { createAuthenticatedContext } from '../auth/apply.js';
 import { RouteInventorySchema, type RouteInventory, type Route } from '../../schemas/route-inventory.schema.js';
 import type { HarnessConfig } from '../../schemas/config.schema.js';
