+++ packages/core/src/tools/scoring/__tests__/gaps.test.ts
@@ -1,6 +1,6 @@
 import test from 'node:test';
 import assert from 'node:assert/strict';
-import { computeCoverageScore, computeQualityScoreFromGaps } from '../gap-engine.js';
+import { computeCoverageScore, computeQualityScoreFromGaps } from '../gaps.js';
 import type { RouteInventory, Route } from '../../../schemas/route-inventory.schema.js';
 import type { Gap } from '../../../schemas/gap-analysis.schema.js';
