+++ packages/core/src/tools/scoring/gaps.ts
